# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agricultural Nursery Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  phone?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  customer: "customer",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Category {
  id: number;
  nameAr: string;
  nameEn: string;
  description?: string | null;
  createdAt: string;
}

export interface CreateCategoryRequest {
  nameAr: string;
  nameEn: string;
  description?: string;
}

export interface Plant {
  id: number;
  nameAr: string;
  nameEn: string;
  categoryId?: number | null;
  category?: Category | null;
  descriptionAr?: string | null;
  descriptionEn?: string | null;
  height?: number | null;
  price: number;
  stockQuantity: number;
  imageUrl?: string | null;
  images: string[];
  featured: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePlantRequest {
  nameAr: string;
  nameEn: string;
  categoryId?: number;
  descriptionAr?: string;
  descriptionEn?: string;
  height?: number;
  price: number;
  stockQuantity: number;
  imageUrl?: string;
  images?: string[];
  featured?: boolean;
}

export interface UpdateStockRequest {
  stockQuantity: number;
}

export interface OrderItem {
  id: number;
  plantId: number;
  plant?: Plant | null;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId?: number | null;
  user?: User | null;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  status: OrderStatus;
  totalAmount: number;
  items: OrderItem[];
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderItemInput {
  plantId: number;
  quantity: number;
}

export interface CreateOrderRequest {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  notes?: string;
  items: OrderItemInput[];
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export interface ReportSummary {
  totalRevenue: number;
  todayRevenue: number;
  monthRevenue: number;
  totalOrders: number;
  todayOrders: number;
  monthOrders: number;
  totalPlants: number;
  lowStockPlants: number;
  totalCustomers: number;
}

export interface DailySales {
  date: string;
  revenue: number;
  orders: number;
}

export interface TopPlant {
  plantId: number;
  nameAr: string;
  nameEn: string;
  totalSold: number;
  totalRevenue: number;
  imageUrl?: string | null;
}

export type GetPlantsParams = {
  search?: string;
  categoryId?: number;
  minPrice?: number;
  maxPrice?: number;
  minHeight?: number;
  maxHeight?: number;
  available?: boolean;
};