# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agricultural Nursery Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from "./orderItemInput";

export interface CreateOrderRequest {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  notes?: string;
  items: OrderItemInput[];
}