# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agricultural Nursery Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";
import type { User } from "./user";

export interface Order {
  id: number;
  userId?: number | null;
  user?: User | null;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  status: OrderStatus;
  totalAmount: number;
  items: OrderItem[];
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}